Split into binary and library
